Create defaults dir and move exitcodes module there.
