+++ salt/client/ssh/__init__.py
@@ -25,7 +25,7 @@ import salt.client.ssh.shell
 import salt.client.ssh.wrapper
 import salt.config
 import salt.exceptions
-import salt.exitcodes
+import salt.defaults.exitcodes
 import salt.log
 import salt.loader
 import salt.minion
@@ -132,7 +132,7 @@ done
 echo "ERROR: Unable to locate appropriate python command" >&2
 exit $EX_PYTHON_OLD
 EOF'''.format(
-    EX_THIN_PYTHON_OLD=salt.exitcodes.EX_THIN_PYTHON_OLD,
+    EX_THIN_PYTHON_OLD=salt.defaults.exitcodes.EX_THIN_PYTHON_OLD,
 )
 
 if not is_windows():
@@ -856,7 +856,7 @@ ARGS = {8}\n'''.format(self.minion_config,
             # RSTR was found in stdout but not stderr - which means there
             # is a SHIM command for the master.
             shim_command = re.split(r'\r?\n', stdout, 1)[0].strip()
-            if 'deploy' == shim_command and retcode == salt.exitcodes.EX_THIN_DEPLOY:
+            if 'deploy' == shim_command and retcode == salt.defaults.exitcodes.EX_THIN_DEPLOY:
                 self.deploy()
                 stdout, stderr, retcode = self.shell.exec_cmd(cmd_str)
                 if not re.search(RSTR_RE, stdout) or not re.search(RSTR_RE, stderr):
@@ -902,12 +902,12 @@ ARGS = {8}\n'''.format(self.minion_config,
                 'sudo expected a password, NOPASSWD required'
             ),
             (
-                (salt.exitcodes.EX_THIN_PYTHON_OLD,),
+                (salt.defaults.exitcodes.EX_THIN_PYTHON_OLD,),
                 'Python interpreter is too old',
                 'salt requires python 2.6 or newer on target hosts'
             ),
             (
-                (salt.exitcodes.EX_THIN_CHECKSUM,),
+                (salt.defaults.exitcodes.EX_THIN_CHECKSUM,),
                 'checksum mismatched',
                 'The salt thin transfer was corrupted'
             ),
