+++ salt/cloud/exceptions.py
@@ -10,7 +10,7 @@
 from __future__ import absolute_import
 
 # Import salt libs
-import salt.exitcodes
+import salt.defaults.exitcodes
 from salt.exceptions import SaltException
 
 
@@ -24,7 +24,7 @@ class SaltCloudSystemExit(SaltCloudException):
     '''
     This exception is raised when the execution should be stopped.
     '''
-    def __init__(self, message, exit_code=salt.exitcodes.EX_GENERIC):
+    def __init__(self, message, exit_code=salt.defaults.exitcodes.EX_GENERIC):
         SaltCloudException.__init__(self, message)
         self.message = '{0} [WARNING: salt.cloud.exceptions is deprecated. Please migrate to salt.exceptions!]'.format(message)
         self.exit_code = exit_code
