+++ salt/defaults/exitcodes.py
